# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FAQ API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/control/local-translations/languages/list": {
            "get": {
                "tags": [
                    "LocalTranslations"
                ],
                "summary": "Get available languages",
                "description": "Returns a list of available languages and versions.",
                "operationId": "getLanguages",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "versions": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    }
                ]
            }
        },
        "/api/control/local-translations/{lang}": {
            "get": {
                "tags": [
                    "LocalTranslations"
                ],
                "summary": "Get translations for a language",
                "description": "Returns translations for the specified language.",
                "operationId": "getTranslations",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "version": {
                                                    "type": "integer"
                                                },
                                                "translations": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Language not found"
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "ApiToken": {
                "type": "apiKey",
                "description": "ApiToken required for all end points",
                "name": "Token",
                "in": "header"
            },
            "SanctumBearerToken": {
                "type": "http",
                "description": "Sanctum Bearer Token required after user logs in (Authorization: Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
...
}